# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BrailleVision API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

/**
 * Input mode for contextual processing
 */
export type BrailleProcessInputMode = typeof BrailleProcessInputMode[keyof typeof BrailleProcessInputMode];


export const BrailleProcessInputMode = {
  camera: 'camera',
  upload: 'upload',
  video: 'video',
} as const;

export interface BrailleProcessInput {
  /** Base64-encoded image data */
  imageBase64: string;
  /** MIME type of the image (image/jpeg, image/png, etc.) */
  mimeType: string;
  /** Input mode for contextual processing */
  mode?: BrailleProcessInputMode;
}

export interface BrailleRegion {
  x: number;
  y: number;
  width: number;
  height: number;
  /** Confidence score 0-1 */
  confidence: number;
}

export interface BrailleProcessResult {
  /** Raw decoded Braille text */
  rawText: string;
  /** Overall confidence score 0-1 */
  confidence: number;
  regions: BrailleRegion[];
  lineCount: number;
  processingMs: number;
  /** Quality warnings (blur, low-contrast, etc.) */
  warnings?: string[];
}

export interface BrailleCorrectionInput {
  /** Raw decoded Braille text to correct */
  rawText: string;
}

export interface BrailleCorrectionResult {
  correctedText: string;
  /** Number of corrections applied */
  changesApplied: number;
}

/**
 * Voice to use for TTS
 */
export type TtsInputVoice = typeof TtsInputVoice[keyof typeof TtsInputVoice];


export const TtsInputVoice = {
  alloy: 'alloy',
  echo: 'echo',
  fable: 'fable',
  onyx: 'onyx',
  nova: 'nova',
  shimmer: 'shimmer',
} as const;

export interface TtsInput {
  text: string;
  /** Voice to use for TTS */
  voice?: TtsInputVoice;
}

export interface TtsResult {
  /** Base64-encoded MP3 audio */
  audioBase64: string;
  format: string;
}

export type ScanMode = typeof ScanMode[keyof typeof ScanMode];


export const ScanMode = {
  camera: 'camera',
  upload: 'upload',
  video: 'video',
} as const;

export interface Scan {
  id: number;
  createdAt: string;
  mode: ScanMode;
  rawText: string;
  correctedText: string;
  confidence: number;
  /** @nullable */
  lineCount?: number | null;
  /** @nullable */
  exportedAt?: string | null;
}

export type ScanInputMode = typeof ScanInputMode[keyof typeof ScanInputMode];


export const ScanInputMode = {
  camera: 'camera',
  upload: 'upload',
  video: 'video',
} as const;

export interface ScanInput {
  mode: ScanInputMode;
  rawText: string;
  correctedText: string;
  confidence: number;
  /** @nullable */
  lineCount?: number | null;
}

export type ScanStatsByMode = {
  camera?: number;
  upload?: number;
  video?: number;
};

export interface ScanStats {
  totalScans: number;
  avgConfidence: number;
  byMode: ScanStatsByMode;
}

export interface ApiError {
  error: string;
}

export type ListScansParams = {
limit?: number;
};