# lib/api-zod/src/generated/types/brailleCorrectionInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BrailleVision API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BrailleCorrectionInput {
  /** Raw decoded Braille text to correct */
  rawText: string;
}