# lib/api-zod/src/generated/types/brailleProcessInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BrailleVision API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BrailleProcessInputMode } from './brailleProcessInputMode';

export interface BrailleProcessInput {
  /** Base64-encoded image data */
  imageBase64: string;
  /** MIME type of the image (image/jpeg, image/png, etc.) */
  mimeType: string;
  /** Input mode for contextual processing */
  mode?: BrailleProcessInputMode;
}