# lib/api-zod/src/generated/types/brailleProcessResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BrailleVision API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BrailleRegion } from './brailleRegion';

export interface BrailleProcessResult {
  /** Raw decoded Braille text */
  rawText: string;
  /** Overall confidence score 0-1 */
  confidence: number;
  regions: BrailleRegion[];
  lineCount: number;
  processingMs: number;
  /** Quality warnings (blur, low-contrast, etc.) */
  warnings?: string[];
}