# lib/api-zod/src/generated/types/brailleRegion.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BrailleVision API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BrailleRegion {
  x: number;
  y: number;
  width: number;
  height: number;
  /** Confidence score 0-1 */
  confidence: number;
}