# lib/api-zod/src/generated/types/scan.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BrailleVision API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ScanMode } from './scanMode';

export interface Scan {
  id: number;
  createdAt: Date;
  mode: ScanMode;
  rawText: string;
  correctedText: string;
  confidence: number;
  /** @nullable */
  lineCount?: number | null;
  /** @nullable */
  exportedAt?: Date | null;
}