# lib/api-zod/src/generated/types/scanInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BrailleVision API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ScanInputMode } from './scanInputMode';

export interface ScanInput {
  mode: ScanInputMode;
  rawText: string;
  correctedText: string;
  confidence: number;
  /** @nullable */
  lineCount?: number | null;
}