# lib/api-zod/src/generated/types/scanStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BrailleVision API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ScanStatsByMode } from './scanStatsByMode';

export interface ScanStats {
  totalScans: number;
  avgConfidence: number;
  byMode: ScanStatsByMode;
}