# lib/api-zod/src/generated/types/ttsInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BrailleVision API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TtsInputVoice } from './ttsInputVoice';

export interface TtsInput {
  text: string;
  /** Voice to use for TTS */
  voice?: TtsInputVoice;
}