# lib/api-zod/src/generated/types/ttsResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * BrailleVision API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TtsResult {
  /** Base64-encoded MP3 audio */
  audioBase64: string;
  format: string;
}